requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
apturl==0.5.2
arrow==1.2.3
async-timeout==4.0.2
attrs==22.2.0
Babel==2.11.0
bcrypt==3.1.7
beautifulsoup4==4.11.1
berserk-downstream==0.11.12
blinker==1.4
Brlapi==0.7.0
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==2.1.1
chrome-gnome-shell==0.0.0
click==8.1.3
colorama==0.4.3
colorclass==2.2.2
command-not-found==0.3
configobj==5.0.6
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
defer==1.0.6
Deprecated==1.2.13
discord.py==2.1.0
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==1.16.0
duplicity==0.8.12.0
entrypoints==0.3
evdev==1.3.0
fasteners==0.14.1
Flask==2.2.2
folder-color-common==0.0.88
frozenlist==1.3.3
future==0.18.2
html5lib==1.0.1
httplib2==0.14.0
ics==0.7.2
idna==3.4
importlib-metadata==5.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
kazam==1.4.5
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.1
matplotlib==3.1.2
monotonic==1.5
multidict==6.0.4
ndjson==0.3.1
netifaces==0.10.4
numpy==1.17.4
oauthlib==3.1.0
olefile==0.46
packaging==23.0
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
pip-upgrader==1.4.15
pony==0.7.16
protobuf==3.6.1
proton-client==0.7.1
protonvpn-cli==3.13.0
protonvpn-gui==1.12.0
protonvpn-nm-lib==3.14.0
psutil==5.5.1
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyRFC3339==1.1
python-apt==2.0.1+zorin1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.21.0
python-gnupg==0.4.5
python-magic==0.4.16
pythondialog==3.4.0
pytz==2022.2.1
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.28.1
requests-unixsocket==0.2.0
SecretStorage==2.3.1
setproctitle==1.1.10
simplejson==3.16.0
six==1.14.0
soupsieve==2.3.2.post1
systemd-python==234
TatSu==5.8.3
terminaltables==3.1.10
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.13
usb-creator==0.3.7
vboxapi==1.0
wadllib==1.3.3
webencodings==0.5.1
Werkzeug==2.2.2
wrapt==1.15.0
xkit==0.0.0
yarl==1.8.2
zorin-appearance==4.1
zorin-connect==1.2
zorin-exec-guard==1.5
